Refactoring of specs, to make them easier to re-create.
