+++ spec/integration/roda/plugins/flow/object_controller_flow_spec.rb
@@ -1,103 +1,10 @@
 require 'spec_helper'
+require 'support/test_repository'
+require 'support/users_controller'
 
 RSpec.describe 'flow plugin' do
   before do
     module Test
-      User = Struct.new(:id, :name, :email)
-
-      class Repository
-        attr_accessor :data
-
-        def initialize(data = {})
-          self.data = data
-        end
-
-        def all
-          data.values
-        end
-
-        def [](id)
-          data[id]
-        end
-
-        def push(obj)
-          obj.id = data.values.length.next
-          data[obj.id] = obj
-        end
-
-        def update(obj)
-          data[obj.id] = obj
-        end
-
-        def delete(obj)
-          data.delete(obj.id)
-        end
-      end
-
-      class UsersController
-        attr_reader :response, :user_repository
-
-        def initialize(response, user_repository)
-          @response = response
-          @user_repository = user_repository
-        end
-
-        def index
-          user_repository.all.map(&:to_h)
-        end
-
-        def show(user_id)
-          if (user = user_repository[user_id.to_i])
-            user.to_h
-          else
-            response.status = 404
-            {}
-          end
-        end
-
-        def create(params)
-          user = User.new(nil, *params.values_at('name', 'email'))
-
-          if !params.values.map(&:to_s).any?(&:empty?)
-            response.status = 201
-            user_repository.push(
-              user
-            ).to_h
-          else
-            response.status = 422
-            user.to_h
-          end
-        end
-
-        def update(user_id, params)
-          if (user = user_repository[user_id.to_i])
-            params = params.each_with_object(user.to_h) { |(k, v), h| h[k.to_sym] = v }
-
-            if params.values.map(&:to_s).any?(&:empty?)
-              response.status = 422
-              params
-            else
-              user_repository.update(
-                User.new(*params.values)
-              ).to_h
-            end
-          else
-            response.status = 404
-            {}
-          end
-        end
-
-        def destroy(user_id)
-          if (user = user_repository[user_id.to_i])
-            user_repository.delete(user)
-            {}
-          else
-            response.status = 404
-            {}
-          end
-        end
-      end
-
       class App < Roda
         plugin :all_verbs
         plugin :json
@@ -137,8 +44,8 @@ RSpec.describe 'flow plugin' do
           end
         end
 
-        register('repositories.user', Repository.new)
-        register('controllers.users') { |*args| UsersController.new(*args) }
+        register('repositories.user', ::TestRepository.new)
+        register('controllers.users') { |*args| ::UsersController.new(*args) }
       end
     end
   end
@@ -154,7 +61,7 @@ RSpec.describe 'flow plugin' do
     before do
       repo = Test::App.resolve('repositories.user')
       users.each do |user_attributes|
-        repo.push(Test::User.new(*user_attributes.values))
+        repo.push(::User.new(*user_attributes.values))
       end
     end
 
@@ -172,10 +79,10 @@ RSpec.describe 'flow plugin' do
     end
 
     before do
-      Test::App.resolve('repositories.user').push(Test::User.new(*user.values))
+      Test::App.resolve('repositories.user').push(::User.new(*user.values))
     end
 
-    context 'with invalid user_id' do
+    context 'with invalid user_id', focus: true do
       it 'returns a 404 with an empty hash representation' do
         get '/users/2', {}
 
@@ -224,7 +131,7 @@ RSpec.describe 'flow plugin' do
     end
 
     before do
-      Test::App.resolve('repositories.user').push(Test::User.new(*user.values))
+      Test::App.resolve('repositories.user').push(::User.new(*user.values))
     end
 
     context 'with invalid user_id' do
@@ -263,7 +170,7 @@ RSpec.describe 'flow plugin' do
     end
 
     before do
-      Test::App.resolve('repositories.user').push(Test::User.new(*user.values))
+      Test::App.resolve('repositories.user').push(::User.new(*user.values))
     end
 
     context 'with invalid user_id' do
