+++ spec/spec_helper.rb
@@ -1,3 +1,4 @@
+$LOAD_PATH.unshift(File.dirname(__FILE__))
 require 'roda'
 require 'rack/test'
