+++ spec/support/test_user.rb
@@ -1,0 +1,1 @@
+User = Struct.new(:id, :name, :email)
